pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nova-brief"
version = "0.1.0"
description = "Deep research agent with citation tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Nova Brief Team" },
]
keywords = ["research", "ai", "citations", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "trafilatura>=1.6.0",
    "pypdf>=3.17.0",
    "tldextract>=5.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/BioInfo/nova-brief"
Repository = "https://github.com/BioInfo/nova-brief"
Documentation = "https://github.com/BioInfo/nova-brief/docs"
Issues = "https://github.com/BioInfo/nova-brief/issues"

[project.scripts]
nova-brief = "src.app:main"
nova-eval = "eval.harness:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]